Refactor: Component 분리
