+++ src/components/mypage/MyProfile.tsx
@@ -1,70 +1,28 @@
-import '@pqina/pintura/pintura.css';
-import { useEffect, useRef, useState } from 'react';
-import { useRecoilState } from 'recoil';
-import { imageState } from '@/atoms/atoms';
-import { PinturaEditorModal } from '@pqina/react-pintura';
-import { getEditorDefaults, createDefaultImageWriter } from '@pqina/pintura';
-import locale_ko_KR from '@pqina/pintura/locale/ko_KR';
-// _PINTURA IMPORTS
+import { useEffect, useState } from 'react';
 import {
   getStorageImg,
   getUserDocRef,
   setProfileImg,
   setStorageImg
 } from '@/api/profile';
-import Icon from '@/components/common/Icon';
+import EditProfileImg from '@/components/mypage/EditProfileImg';
 import CheckNickname from '@/components/start/CheckNickname';
 import { TEXT } from '@/constants/texts';
 import { useComposeHeader } from '@/hooks/useComposeHeader';
-import { iconPropsGenerator } from '@/utils/iconPropsGenerator';
-import { FlexCenter, Justify, Column } from '@/styles/layout';
+import { FlexCenter, Justify } from '@/styles/layout';
 import { Cursor, LineH18, TextGray, Border16, Medium } from '@/styles/styles';
 import { styled } from 'styled-system/jsx';
 import { cx } from 'styled-system/css';
 import { CachedData } from '@/types/types';
 import useGetCacheData from '@/hooks/useGetCacheData';
-
-const editorDefaults = getEditorDefaults({
-  cropImageSelectionCornerStyle: 'hook',
-  locale: {
-    ...locale_ko_KR
-  },
-  imageWriter: createDefaultImageWriter({
-    targetSize: {
-      width: 100,
-      height: 100,
-      fit: 'contain',
-      upscale: true
-    }
-  })
-});
+import { useRecoilState } from 'recoil';
+import { imageState } from '@/atoms/atoms';
 
 const MyProfile = () => {
   useComposeHeader(false, '프로필 수정', 'close');
 
   const handleExitedUser = () => {
     console.log('회원 탈퇴');
-  };
-
-  const [editorEnabled, setEditorEnabled] = useState(false);
-  const [editorSrc, setEditorSrc] = useState<File>();
-  const [imageUrl, setImageUrl] = useRecoilState(imageState);
-
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const handleInputChange = () => {
-    if (!fileInputRef?.current?.files?.length) return;
-    if (fileInputRef?.current?.files) {
-      setEditorEnabled(true);
-      setEditorSrc(fileInputRef?.current?.files[0]);
-    }
-  };
-
-  const handleEditorHide = () => setEditorEnabled(false);
-
-  const handleEditorProcess = ({ dest }: { dest: File }) => {
-    const url = URL.createObjectURL(dest);
-    setImageUrl(url);
   };
 
   const [cachedData, setCachedData] = useState<CachedData>();
@@ -80,91 +38,42 @@ const MyProfile = () => {
     getCachedUserInfo();
   }, []);
 
-  const handleFormSubmit = async () => {
-    if (userId) {
-      const fileInput = fileInputRef.current;
+  const [imageUrl, setImageUrl] = useRecoilState(imageState);
 
-      if (fileInput?.files?.length) {
-        const file = fileInput.files[0];
+  const handleFormSubmit = async () => {
+    if (userId && imageUrl) {
+      try {
         const userDocRef = await getUserDocRef();
         const filePath = `users/${userId}/profileImage.jpg`;
+
+        const response = await fetch(imageUrl);
+        const blob = await response.blob();
+
+        const file = new File([blob], 'profileImage.jpg', {
+          type: 'image/jpeg'
+        });
         await setStorageImg(filePath, file);
         await getStorageImg(filePath);
         await setProfileImg(userDocRef, filePath);
+      } catch (err) {
+        console.log(err);
       }
     }
   };
+
+  const handleImageSelect = (selectedImage: File) => {
+    setImageUrl(URL.createObjectURL(selectedImage));
+  };
+
   return (
     <>
-      <Wrapper className={cx(FlexCenter, Column)}>
-        <ImgContainer>
-          {imageUrl && (
-            <ImgRound>
-              <img
-                src={imageUrl}
-                alt="profile image"
-              />
-            </ImgRound>
-          )}
-          {!imageUrl && <Icon {...iconPropsGenerator('user', '100')} />}
-          <Edit className={Cursor}>
-            <label>
-              <Icon {...iconPropsGenerator('edit-photo', '32')} />
-              <input
-                ref={fileInputRef}
-                type="file"
-                accept="image/*"
-                onChange={handleInputChange}
-                style={{ display: 'none' }}
-              />
-            </label>
-          </Edit>
-        </ImgContainer>
-      </Wrapper>
+      <EditProfileImg onImageSelect={handleImageSelect} />
       <CheckNickname />
-
       <ExitButton
         className={cx(Cursor, LineH18, TextGray)}
         onTouchEnd={handleExitedUser}>
         {TEXT.exitButtonText}
       </ExitButton>
-
-      {editorEnabled && (
-        <PinturaEditorModal
-          {...editorDefaults}
-          src={editorSrc}
-          imageCropAspectRatio={1}
-          onHide={handleEditorHide}
-          onProcess={handleEditorProcess}
-          willRenderCanvas={(shapes, state) => {
-            const { utilVisibility, selectionRect, lineColor } = state;
-
-            if (utilVisibility.crop <= 0) return shapes;
-
-            const { x, y, width, height } = selectionRect;
-
-            return {
-              ...shapes,
-
-              interfaceShapes: [
-                {
-                  x: x + width * 0.5,
-                  y: y + height * 0.5,
-                  rx: width * 0.5,
-                  ry: height * 0.5,
-                  opacity: utilVisibility.crop,
-                  inverted: true,
-                  backgroundColor: [0, 0, 0, 0.1],
-                  strokeWidth: 0.4,
-                  strokeColor: [...lineColor]
-                },
-                ...shapes.interfaceShapes
-              ]
-            };
-          }}
-        />
-      )}
-
       <ButtonArea className={Justify}>
         <SaveButton
           className={cx(FlexCenter, Cursor, Border16, Medium)}
@@ -176,19 +85,6 @@ const MyProfile = () => {
   );
 };
 
-const Wrapper = styled.div`
-  padding: 20px 0 40px;
-  gap: 40px;
-`;
-
-const Edit = styled.div`
-  position: absolute;
-  width: 30px;
-  height: 30px;
-  bottom: 9px;
-  right: -2px;
-  z-index: 1;
-`;
 const ExitButton = styled.span`
   font-size: var(--font--sizes-sm);
   padding: 16px 0;
@@ -206,15 +102,5 @@ const SaveButton = styled.button`
   background-color: var(--colors-main);
   font-size: var(--font-sizes-base);
   color: #fff;
-`;
-
-const ImgContainer = styled.div`
-  position: relative;
-`;
-const ImgRound = styled.div`
-  width: 100px;
-  height: 100px;
-  border-radius: 100px;
-  overflow: hidden;
 `;
 export default MyProfile;
